requirements.txt
pandas
numpy
setuptools
ipykernel
google-cloud-storage
pyyaml
matplotlib
tensorflow
wordcloud
scikit-learn
comet-ml
dotenv
dvc
## C:\Users\Sagar\Downloads\ace-wording-455116-q0-d134b6ddc061.json